requirements.txt
APScheduler==3.6.3
autopep8==1.6.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Flask==2.1.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pycodestyle==2.8.0
pyTelegramBotAPI==4.5.0
python-dotenv==0.20.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
toml==0.10.2
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
